third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/nonces/scriptnonce-allowed.html

<!DOCTYPE html>
<html>
    <head>
        <meta http-equiv="Content-Security-Policy" content="script-src 'nonce-noncynonce' 'nonce-noncy+/='">
        <script nonce="noncynonce">
        if (window.testRunner)
            testRunner.dumpAsText();
        </script>
        <script nonce="noncynonce">
            alert('PASS (1/2)');
        </script>
        <script nonce="noncy+/=">
            alert('PASS (2/2)');
        </script>
    </head>
    <body>
        <p>
            This tests the effect of a valid script-nonce value. It passes if
            no console warning is visible and the alerts are executed.
        </p>
    </body>
</html>
